refactor(FloatingBox): Use `partials` pattern
